# src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZeroAuth API",
    "version": "1.0.0",
    "description": "Zero-trust authentication and authorization API. Built on PASETO v4, ABAC/RBAC, device attestation, and continuous access evaluation.",
    "contact": {
      "email": "mas.arafat.dev@gmail.com"
    }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local development" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "PASETO"
      }
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "required": ["code", "message", "details"],
        "properties": {
          "code": { "type": "string" },
          "message": { "type": "string" },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": { "type": "string" },
                "message": { "type": "string" }
              }
            }
          }
        }
      },
...
      "Session": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "ipAddress": { "type": "string" },
          "country": { "type": "string" },
          "userAgent": { "type": "string" },
          "isActive": { "type": "boolean" },
          "isCurrent": { "type": "boolean" },
          "lastActivityAt": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" },
          "deviceFingerprint": {
            "type": "object",
            "properties": {
              "platform": { "type": "string" },
              "browser": { "type": "string" },
              "os": { "type": "string" }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/sessions": {
      "get": {
        "tags": ["Sessions"],
        "summary": "List active sessions for authenticated user",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "activeOnly", "in": "query", "schema": { "type": "boolean", "default": true } }
        ],
        "responses": {
          "200": {
            "description": "Session list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sessions": { "type": "array", "items": { "$ref": "#/components/schemas/Session" } },
                    "total": { "type": "integer" },
                    "limit": { "type": "integer" },
                    "offset": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Sessions"],
        "summary": "Revoke all other sessions (keep current)",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Sessions revoked", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "revokedCount": { "type": "integer" } } } } } }
        }
      }
    },
...
    "/ssf/events": {
      "post": {
        "tags": ["Shared Signals"],
        "summary": "Receive a Security Event Token (SET) from a provider",
        "parameters": [
          { "name": "x-ssf-signature", "in": "header", "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "description": "RFC 8417 Security Event Token payload" } } }
        },
        "responses": {
          "200": { "description": "Event handled" },
          "401": { "description": "Invalid SSF signature", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } }
        }
      }
    },
...
  }
}